pyproject.toml
[tool.poetry]
name = "research"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.26.4"
opencv-python = "^4.10.0.84"
pandas = "^2.2.2"
matplotlib = "^3.9.1.post1"
seaborn = "^0.13.2"
plotly = "^5.23.0"
scikit-image = "^0.24.0"
pillow = "^10.4.0"
notebook = "^7.2.1"
streamlit = "^1.37.1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
